requirements.txt
pydub==0.25.1
streamlit==1.35.0
transformers==4.41.0
sentencepiece==0.2.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab